requirements.txt
streamlit
python-dotenv
groq
langchain-community
langchain
requests
PyPDF2
numpy
